Refactor dct references
