+++ .rubocop.yml
@@ -84,6 +84,12 @@ RSpec/VerifiedDoubles:
     - 'spec/lib/geoblacklight/download/hgl_download_spec.rb'
     - 'spec/lib/geoblacklight/document_presenter_spec.rb'
 
+Style/MethodMissing:
+  Exclude:
+    - 'app/models/concerns/geoblacklight/solr_document.rb'
+    - 'lib/geoblacklight/dct_references.rb'
+    - 'lib/geoblacklight/references.rb'
+
 Style/SignalException:
   EnforcedStyle: semantic
