+++ lib/geoblacklight.rb
@@ -27,7 +27,9 @@ module Geoblacklight
   require 'geoblacklight/metadata_transformer/base'
   require 'geoblacklight/metadata_transformer/fgdc'
   require 'geoblacklight/metadata_transformer/iso19139'
+  require 'geoblacklight/dct_reference'
   require 'geoblacklight/reference'
+  require 'geoblacklight/dct_references'
   require 'geoblacklight/references'
   require 'geoblacklight/routes'
   require 'geoblacklight/relation/descendants'
