+++ lib/geoblacklight/dct_reference.rb
@@ -1,0 +1,45 @@
+module Geoblacklight
+  ##
+  # Parses an array of dct_references to create useful reference information
+  class DctReference
+    attr_reader :reference
+
+    ##
+    # Initializes a DctReference object using an Array
+    # @param [Array] reference
+    def initialize(reference)
+      @reference = reference
+    end
+
+    ##
+    # The endpoint URL for a Geoblacklight::DctReference
+    # @return [String]
+    def endpoint
+      @reference[1]
+    end
+
+    ##
+    # Lookups the type from the Constants::URI using the reference's URI
+    # @return [Symbol]
+    def type
+      Geoblacklight::Constants::URI.key(uri)
+    end
+
+    ##
+    # Creates a hash, using its type as key and endpoint as value
+    # @return [Hash]
+    def to_hash
+      { type => endpoint }
+    end
+
+    private
+
+      ##
+      # The URI used for this instance's creation
+      # Remove any trailing slashes
+      # @return [String]
+      def uri
+        @reference[0].sub(/(\/)+$/, '') if @reference[0].present?
+      end
+  end
+end
