+++ lib/geoblacklight/item_viewer.rb
@@ -14,41 +14,43 @@ module Geoblacklight
       viewer_preference.values.first.to_s
     end
 
-    def wms
-      @references.wms
-    end
+    private
 
-    def iiif
-      @references.iiif
-    end
+      def wms
+        @references.wms
+      end
 
-    def tiled_map_layer
-      @references.tiled_map_layer
-    end
+      def iiif
+        @references.iiif
+      end
 
-    def dynamic_map_layer
-      @references.dynamic_map_layer
-    end
+      def tiled_map_layer
+        @references.tiled_map_layer
+      end
 
-    def feature_layer
-      @references.feature_layer
-    end
+      def dynamic_map_layer
+        @references.dynamic_map_layer
+      end
 
-    def image_map_layer
-      @references.image_map_layer
-    end
+      def feature_layer
+        @references.feature_layer
+      end
 
-    def index_map
-      @references.index_map
-    end
+      def image_map_layer
+        @references.image_map_layer
+      end
 
-    def oembed
-      @references.oembed
-    end
+      def index_map
+        @references.index_map
+      end
 
-    def viewer_preference
-      [oembed, index_map, wms, iiif, tiled_map_layer, dynamic_map_layer,
-       image_map_layer, feature_layer].compact.map(&:to_hash).first
-    end
+      def oembed
+        @references.oembed
+      end
+
+      def viewer_preference
+        [oembed, index_map, wms, iiif, tiled_map_layer, dynamic_map_layer,
+         image_map_layer, feature_layer].compact.map(&:to_hash).first
+      end
   end
 end
