+++ spec/lib/geoblacklight/metadata/html_spec.rb
@@ -4,7 +4,7 @@ describe Geoblacklight::Metadata::Html do
   let(:url) { 'https://s3.amazonaws.com/cugir-data/00/77/41/fgdc.html' }
   let(:metadata) do
     described_class.new(
-      Geoblacklight::Reference.new(
+      Geoblacklight::DctReference.new(
         ['http://www.w3.org/1999/xhtml', url]
       )
     )
